+++ src/components/AboutMe/AboutMe.jsx
@@ -14,9 +14,10 @@ export default function AboutMe() {
             <h3 className='aboutme__name'>Илья</h3>
             <p className='aboutme__subheader'>Фронтенд-разработчик, 33 года</p>
             <p className='aboutme__paragraph'>
-              Я родился и живу в Саратове, закончил факультет экономики СГУ. У меня есть жена и дочь. Я люблю слушать
-              музыку, а ещё увлекаюсь бегом. Недавно начал кодить. С 2015 года работал в компании «СКБ Контур». После
-              того, как прошёл курс по веб-разработке, начал заниматься фриланс-заказами и ушёл с постоянной работы.
+              Я родился Москве, а теперь живу на солнечном побережье Турции, закончил факультет химической технологии
+              РХТУ им. Д.И.Менделеева. Мне нравится путешествовать, скалолазанье, сноуборд и научпоп. Недавно начал
+              кодить. С 2022 года работаю в компании «Flying Academy» фронтенд-разработчиком, а также помогаю делать
+              функционал для одного приложения в качестве фрилансера.
             </p>
             <a className='aboutme__link' href='https://github.com/iluxmas' target='_blank'>
               Github
