+++ src/components/Logo/Logo.css
@@ -6,4 +6,8 @@
   background-size: 38px 38px;
   padding-left: 38px;
   height: 38px;
+  transition: opacity 0.2s ease;
+}
+.logo:hover {
+  opacity: 0.6;
 }
