+++ src/components/Logo/Logo.jsx
@@ -4,5 +4,5 @@ import { NavLink } from 'react-router-dom';
 import './Logo.css';
 
 export default function Logo() {
-  return <NavLink className='logo' to='/'></NavLink>;
+  return <NavLink className='logo' to='/' title='На главную'></NavLink>;
 }
