+++ src/components/MoviesCard/MoviesCard.css
