+++ src/components/Preloader/Preloader.jsx
@@ -5,7 +5,9 @@ import './Preloader.css';
 export default function Preloader() {
   return (
     <div className='preloader'>
-      <button className='preloader__button'>Ещё</button>
+      <div className='preloader__container'>
+        <span className='preloader__round'></span>
+      </div>
     </div>
   );
 }
