+++ src/components/SavedMovies/SavedMovies.jsx
@@ -1,39 +1,25 @@
 import React from 'react';
 import SearchForm from '../SearchForm/SearchForm';
 import MoviesCardList from '../MoviesCardList/MoviesCardList';
-import MovieCard from '../MovieCard/MovieCard';
+import MoviesCard from '../MoviesCard/MoviesCard';
 import './SavedMovies.css';
 
 export default function SavedMovies() {
+  const cards = Array(4)
+    .fill(1)
+    .map((a, idx) => (
+      <MoviesCard
+        title={'Mesto mesto mesto'}
+        duration={122}
+        key={idx}
+        posterURL={'https://github.com/Iluxmas/resume/raw/master/src/images/projects/mesto-app.JPG'}
+        isFav={2 > 1}
+      />
+    ));
   return (
     <section className='saved-movies'>
       <SearchForm />
-      <MoviesCardList>
-        <MovieCard
-          title={'alabai'}
-          duration={122}
-          posterURL={'https://github.com/Iluxmas/resume/raw/master/src/images/projects/mesto-app.JPG'}
-          isFav={false}
-        />
-        <MovieCard
-          title={'alabai'}
-          duration={122}
-          posterURL={'https://github.com/Iluxmas/resume/raw/master/src/images/projects/mesto-app.JPG'}
-          isFav={false}
-        />
-        <MovieCard
-          title={'alabai'}
-          duration={122}
-          posterURL={'https://github.com/Iluxmas/resume/raw/master/src/images/projects/mesto-app.JPG'}
-          isFav={false}
-        />
-        <MovieCard
-          title={'alabai'}
-          duration={122}
-          posterURL={'https://github.com/Iluxmas/resume/raw/master/src/images/projects/mesto-app.JPG'}
-          isFav={false}
-        />
-      </MoviesCardList>
+      <MoviesCardList>{cards}</MoviesCardList>
     </section>
   );
 }
