+++ src/components/ShowMore/ShowMore.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+import './ShowMore.css';
+
+export default function ShowMore() {
+  return (
+    <div className='show-more'>
+      <button className='show-more__button'>Ещё</button>
+    </div>
+  );
+}
